+++ lib/cce.rb
@@ -7,7 +7,7 @@
 #   - Own no other logic itself.
 
 module CCE
-  VERSION = "2.4.0"
+  VERSION = "2.4.1"
 
   @registry = nil